fix: Move SDNClient from utils to api_client directory and dedicated file
